+++ README.md
@@ -1,53 +1,13 @@
-# Sobriety
-
-## What Is Sobriety?
-
-Sobriety is a simple and minimal app to track how long you've been sober from anything you want. It'll tell you how long it's been since you logged a relapse, and the average time between your three most recent relapses, so that you have more input with which to work on recovery. Sobriety has no trackers, and runs completely on your device, and every feature is completely free.
-
-[<img src="https://fdroid.gitlab.io/artwork/badge/get-it-on.png"
-     alt="Get it on F-Droid"
-     height="80">](https://f-droid.org/packages/com.katiearose.sobriety/)
-
-or get the APK from the [Releases section](https://github.com/KiARC/Sobriety/releases/latest).
-
-## Why Does Sobriety Exist?
-
-I mostly designed this out of spite for I Am Sober, which monetizes addiction recovery. This started as something I just wrote in the time I had between classes for a friend who was struggling with addiction and couldn't get an I Am Sober subscription to track all the things he was struggling with, and I really just decided to publish it because I didn't see a reason not to, but then it got picked up by the IzzyOnDroid repository and I decided it was genuinely going to be worth working on further if people were actually aware of it. Since then a lot of tweaks have been made and the amazing JakeGonder has helped out quite a bit, and I am honestly super proud of how far this has come. Recently (as of writing this),  Donnnno, the creator of [the impressive Arcticons icon pack](https://github.com/Donnnno/Arcticons) has added a Sobriety icon, and I am over the moon with pride for the recognition. I hope Sobriety is able to help more people, and I will continue to work on it when I'm able to because I want to see it grow.
-
-### About I Am Sober
-
-I'm only half joking when I say I hate I Am Sober. [It has several trackers and requires far more permissions than is reasonable for the features it provides](https://reports.exodus-privacy.eu.org/en/reports/com.thehungrywasp.iamsober/latest/), and the fact that a subscription is required to track more than two individual addictions is just horrific to me. Displaying the time since a given date, and allowing that date to be reset to the current time, are such simple features to implement that I genuinely cannot fathom how they can in good conscience monetize them, especially when their target audience is *people struggling with addiction*. I just find the whole concept somewhat upsetting. For that reason, Sobriety will never have any trackers or anything of the sort, and will always be completely free. I want to help people, not profit off of their struggles.
-
-## Developer Contact Info
-- Here on GitHub, via Issues, Discussions, etc
-- My personal email address (katiearose@pm.me)
-
-Feel free to contact me about any issues, suggestions or comments you have about Sobriety!
-
-# Repository Metadata (via Shields.io)
-## Core
-![GitHub](https://img.shields.io/github/license/KiARC/Sobriety?style=for-the-badge?)
-![GitHub Workflow Status](https://img.shields.io/github/workflow/status/KiARC/Sobriety/Android%20CI?label=CI&style=for-the-badge?)
-![Maintenance](https://img.shields.io/maintenance/yes/2022?style=for-the-badge?)
-## Size
-![GitHub code size](https://img.shields.io/github/languages/code-size/KiARC/Sobriety?style=for-the-badge?)
-![GitHub repo size](https://img.shields.io/github/repo-size/KiARC/Sobriety?style=for-the-badge?)
-## Release Info
-![GitHub all releases](https://img.shields.io/github/downloads/KiARC/Sobriety/total?style=for-the-badge&label=Downloads%20%28All%20Releases%29?)
-![GitHub downloads by release (latest by date)](https://img.shields.io/github/downloads/KiARC/Sobriety/latest/total?style=for-the-badge&label=Downloads%20%28Latest%20Release%29?)
-
-![GitHub release (latest SemVer)](https://img.shields.io/github/v/release/KiARC/Sobriety?label=Latest%20Release&sort=semver&style=for-the-badge?)
-![GitHub Release Date](https://img.shields.io/github/release-date/KiARC/Sobriety?label=Latest%20Release%20Date&style=for-the-badge&sort=semver?)
-![IzzyOnDroid](https://img.shields.io/endpoint?url=https://apt.izzysoft.de/fdroid/api/v1/shield/com.katiearose.sobriety&label=IzzyOnDroid&style=for-the-badge?)
-## Issue Counts
-![GitHub open issues](https://img.shields.io/github/issues-raw/KiARC/Sobriety?style=for-the-badge?)
-![GitHub closed issues](https://img.shields.io/github/issues-closed-raw/KiARC/Sobriety?style=for-the-badge?)
-## PR Counts
-![GitHub open pull requests](https://img.shields.io/github/issues-pr-raw/KiARC/Sobriety?style=for-the-badge?)
-![GitHub closed pull requests](https://img.shields.io/github/issues-pr-closed-raw/KiARC/Sobriety?style=for-the-badge?)
-## Stats
-![GitHub commit activity](https://img.shields.io/github/commit-activity/w/KiARC/Sobriety?style=for-the-badge?)
-![GitHub commits since latest release (by date)](https://img.shields.io/github/commits-since/KiARC/Sobriety/latest?style=for-the-badge?)
-
-![GitHub contributors](https://img.shields.io/github/contributors/KiARC/Sobriety?style=for-the-badge?)
-![GitHub Discussions](https://img.shields.io/github/discussions/KiARC/Sobriety?style=for-the-badge?)
+# I'm Definitely Sober
+## Why does this branch exist?
+Because [the upstream developer](https://github.com/KiARC) can't seem to keep up with the speed at which I'm developing the app, so I decided to spin it off.
+## What's the difference between this version and the upstream?
+This version has much faster development speed (well, most of the time). That means features and bug fixes are more frequent, but at the same time, new bugs and regressions can happen. (Sort of like the *nightly* version of some apps). **If you find any, please open an issue.**
+## Why is it distinct from [the upstream app?](https://github.com/KiARC/Sobriety)
+Two reasons:
+- I want to have both my version and the upstream version installed
+- In the worst case scenario, I can publish my version on [F-Droid](https://f-droid.org) later.
+## Do you plan to contribute to the upstream version later?
+Of course! In fact, the only reason why this version exists is because the upstream dev isn't fast enough to respond to my contributions!
+## Final notes
+Katherine, if you're reading this... Thank you for your great contribution to the world. If you want to ask me anything, feel free to open an issue.
