+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.katiearose.sobriety">
+    package="com.sixtyninefourtwenty.imdefinitelysober">
 
     <application
         android:allowBackup="true"
@@ -17,8 +17,7 @@
             android:windowSoftInputMode="adjustResize" />
         <activity
             android:name=".Main"
-            android:exported="true"
-            android:label="@string/app_name">
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
