+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">Sobriety</string>
+    <string name="app_name">I\'m Definitely Sober</string>
     <string name="title_activity_main">Main</string>
     <string name="large_text">
         "Material is the metaphor.\n\n"
